+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - (Bugfix) (DP) Propagate Timeout Across Subcommands
 - (Maintenance) Bump Dependencies
 - (Feature) (Platform) EventsV1 Integration
+- (Feature) (Platform) Allows to opt out in the Inventory Telemetry 
 
 ## [1.3.1](https://github.com/arangodb/kube-arangodb/tree/1.3.1) (2025-10-07)
 - (Documentation) Add ArangoPlatformStorage Docs & Examples